+++ source_gen/source_gen/lib/src/generator.dart
@@ -4,9 +4,11 @@
 
 import 'dart:async';
 
+import 'package:analyzer/dart/element/element.dart';
 import 'package:build/build.dart';
 
 import 'library.dart';
+import 'span_for_element.dart';
 
 /// A tool to generate Dart code based on a Dart library source.
 ///
@@ -36,9 +38,25 @@ class InvalidGenerationSourceError extends Error {
   /// May be an empty string if unknown.
   final String todo;
 
-  InvalidGenerationSourceError(this.message, {String todo})
+  /// The code element associated with this error.
+  ///
+  /// May be `null` if the error had no associated element.
+  final Element element;
+
+  InvalidGenerationSourceError(this.message, {String todo, this.element})
       : this.todo = todo ?? '';
 
   @override
-  String toString() => message;
+  String toString() {
+    var buffer = new StringBuffer(message);
+
+    if (element != null) {
+      var span = spanForElement(element);
+      buffer.writeln();
+      buffer.writeln(span.start.toolString);
+      buffer.write(span.highlight());
+    }
+
+    return buffer.toString();
+  }
 }