+++ source_gen/source_gen/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.8.1
+
+* `InvalidGenerationSourceError` added an optional `element`
+  parameter to support more helpful error messages.
+
 ## 0.8.0
 
 * **BREAKING** removed the deprecated `requireLibraryDirective` parameter in